added heartbeat message to maintain link base sends this out to all endpoints repeatedly after a link change, if it does not recieve a response, it reverts its radio settings
After recieving a linkChangeRsp, instead of stopping linkChangeTimeoutTimer, serviceLinkChangeRsp resets it and extends the timeout using the new setTimeout function.
linkChangeTimeoutTimer still calls linkChangeReqTimeout to revert settings when it times out.
successfully recieving a message resets linkChangeReqTimeout, and if SUCCESSFUL_PACKETS_BEFORE_LINK_IS_TRUSTED packets or acknowlegements are recieved, it trusts the link and stops the timer.